first attempt at a postMessage() handshake
At this early stage, this feels like it might be too much ceremony. I
am anticipating need to migrate protocols for the handshake and for
storage.

However, part of the conversation I hope to start with my
collaborators is precisely how to set up this first version so that we
have room to grow. Thinking especially of two decisions in federated
wiki that we wish were different:

1. the structure of lineup URLs requires a server to interpret the GET
requests, or requires gymnastics using 404.html conventions

2. the name of wiki files on disk excludes a file extension which
again requires a server to provide the appropriate mime type for
browsers.