added example clip tool and bookmarklet
The bookmarklet when installed allows the active reader to save a
bookmark with a bit more information than the usual workflow. We are
interested to record why we care about saving this link.

The app captures the data sent from the bookmarklet into browser local
storage. It also has a light interface to display the items captured
and allow the reader to clear all the data.