overhaul clip.html and clip.js
There are now two different state machines in this code. The top level
state handles the two contexts currently implemented: saving data sent
from the bookmarklet, and showing what's currently saved.

In the "show" state we have an internal state machine for the buttons
that allow the active reader to clear the collection of saved clips.