added wiki showResult behavior
This might work as-is for any wiki, provided the clip.html and .js are
hosted in an assets folder for that wiki. I might be able to
re-purpose code from ./welcome-active-reader.js to make it work across
domains, but not positive that's worth the effort when it is pretty
easy to fork assets to other wikis.